// src/types/tech.ts
// Interface automatically generated by schemas-to-ts

import { Project } from '@/types/project';
import { Project_Plain } from '@/types//project';
import { AdminPanelRelationPropertyModification } from '@/types//AdminPanelRelationPropertyModification';

export interface Tech {
  id: number;
  attributes: {
    createdAt: Date;
    updatedAt: Date;
    publishedAt?: Date;
    Tech: string;
    projects?: { data: Project[] };
    Slug: string;
  };
}
export interface Tech_Plain {
  id: number;
  createdAt: Date;
  updatedAt: Date;
  publishedAt?: Date;
  Tech: string;
  projects?: Project_Plain[];
  Slug: string;
}

export interface Tech_NoRelations {
  id: number;
  createdAt: Date;
  updatedAt: Date;
  publishedAt?: Date;
  Tech: string;
  projects?: number[];
  Slug: string;
}

export interface Tech_AdminPanelLifeCycle {
  id: number;
  createdAt: Date;
  updatedAt: Date;
  publishedAt?: Date;
  Tech: string;
  projects?: AdminPanelRelationPropertyModification<Project_Plain>;
  Slug: string;
}